fix cython build (#6356)
* fix cython

* fix
