+++ python/mxnet/cython/symbol.pyx
@@ -192,10 +192,15 @@ def _init_symbol_module(symbol_class, root_namespace):
 
     module_obj = _sys.modules["%s.symbol" % root_namespace]
     module_internal = _sys.modules["%s._symbol_internal" % root_namespace]
+    module_contrib = _sys.modules["%s.contrib.symbol" % root_namespace]
     for i in range(op_names.size()):
         CALL(NNGetOpHandle(op_names[i].c_str(), &handle))
         function = _make_atomic_symbol_function(handle, op_names[i])
-        if function.__name__.startswith('_'):
+        if function.__name__.startswith('_contrib_'):
+            function.__name__ = function.__name__[9:]
+            function.__module__ = 'mxnet.contrib.symbol'
+            setattr(module_contrib, function.__name__, function)
+        elif function.__name__.startswith('_'):
             setattr(module_internal, function.__name__, function)
         else:
             setattr(module_obj, function.__name__, function)
