+++ python/setup.py
@@ -4,12 +4,18 @@ from __future__ import absolute_import
 import os
 import sys
 # need to use distutils.core for correct placement of cython dll
+kwargs = {}
 if "--inplace" in sys.argv:
     from distutils.core import setup
     from distutils.extension import Extension
 else:
     from setuptools import setup
     from setuptools.extension import Extension
+    kwargs = {'install_requires': ['numpy'], 'zip_safe': False}
+
+if '--with-cython' in sys.argv:
+    with_cython = True
+    sys.argv.remove('--with-cython')
 
 # We can not import `mxnet.info.py` in setup.py directly since mxnet/__init__.py
 # Will be invoked which introduces dependences
@@ -24,7 +30,8 @@ __version__ = libinfo['__version__']
 
 def config_cython():
     """Try to configure cython and return cython configuration"""
-    return [] # disable cython due to some users have compile errors.
+    if not with_cython:
+        return []
     # pylint: disable=unreachable
     if os.name == 'nt':
         print("WARNING: Cython is not supported on Windows, will compile without cython module")
@@ -65,14 +72,11 @@ def config_cython():
 setup(name='mxnet',
       version=__version__,
       description=open(os.path.join(CURRENT_DIR, 'README.md')).read(),
-      install_requires=[
-          'numpy',
-      ],
-      zip_safe=False,
       packages=[
           'mxnet', 'mxnet.module', 'mxnet._ctypes', 'mxnet.rnn',
           'mxnet._cy2', 'mxnet._cy3', 'mxnet.notebook', 'mxnet.contrib'
           ],
       data_files=[('mxnet', [LIB_PATH[0]])],
       url='https://github.com/dmlc/mxnet',
-      ext_modules=config_cython())
+      ext_modules=config_cython(),
+      **kwargs)
